requirements.txt
astroid==2.5.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
inky==1.2.0
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.20.1
Pillow==8.1.2
PyJWT==1.7.1
pylint==2.7.2
python-dotenv==0.15.0
pytz==2021.1
ratelimit==2.2.1
requests==2.25.1
six==1.15.0
smbus2==0.4.1
spidev==3.5
toml==0.10.2
twilio==6.54.0
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1

